modify test and delete unnecessary files
